Fix anime enums #JC19-326
