+++ libs/core/models/anime.ts
@@ -1,22 +1,7 @@
+import { AnimeEnums } from '../utils/types/animeEnums';
+
 import { DateRange } from './dateRange';
 import { Immerable, OmitImmerable } from './immerable';
-
-/** Possible options anime type. */
-export enum Type {
-  Tv = 'TV',
-  Ova = 'OVA',
-  Movie = 'MOVIE',
-  Special = 'SPECIAL',
-  Ona = 'ONA',
-  Music = 'MUSIC',
-}
-
-/** Possible options anime status. */
-export enum Status {
-  Airing = 'AIRING',
-  Finished = 'FINISHED',
-  NotYetAired = 'NOT_YET_AIRED',
-}
 
 /** Anime. */
 export class Anime extends Immerable {
@@ -37,10 +22,10 @@ export class Anime extends Immerable {
   public readonly aired: DateRange;
 
   /** Anime type. */
-  public readonly type: Type;
+  public readonly type: AnimeEnums.Type;
 
   /** Anime status. */
-  public readonly status: Status;
+  public readonly status: AnimeEnums.Status;
 
   public constructor(data: InitArgsAnime) {
     super();
