+++ libs/core/utils/types/animeEnums.ts
@@ -1,0 +1,20 @@
+/** All enums relating to anime. */
+export namespace AnimeEnums {
+
+  /** Possible options anime type. */
+  export enum Type {
+    Tv = 'TV',
+    Ova = 'OVA',
+    Movie = 'MOVIE',
+    Special = 'SPECIAL',
+    Ona = 'ONA',
+    Music = 'MUSIC',
+  }
+
+  /** Possible options anime status. */
+  export enum Status {
+    Airing = 'AIRING',
+    Finished = 'FINISHED',
+    NotYetAired = 'NOT_YET_AIRED',
+  }
+}
